feat(project): Add button component, update components to use new tailwind config
A number of changes have been made to:
I have modified the `Button` and styled it to look like the GovUK
component. This includes all states `hover`, `active`, `focus`, etc.
I have created new stories and tests for this new button.

Due to the changes in the tailwind config and the names of colours
and other properties, I have updated the existing components and
tests to use these new names.

I have added `tailwind-merge` as a dependency, so components can
optionally take extra user supplied class names and they will be
applied correct. `react-icons` v5.4.0 has been added as a peer
dependency.

To explain the components and some of the design decisions I have
updated the readme as well.