feat(deps): Update outdated dependencies
Some of the main dependencies like `react` were
a major version behind and needed updating.
Most of the upgrades were simple and did not cause any issues,
but some packages were deprecated/replaced/had breaking changes
that I will list below:

- `@storybook/addon-styling`: this has been divided into
2 libraries, one with `webpack` support that we dont' want,
 and the other `@storybook/addon-themes` which adds the theme
behaviour.
- `@storybook/testing-library`: this has been deprecated in
favour of `@storybook/test` which has builtin `vitest` support.
- `eslint-plugin-vitest`: although not explicitly deprecated,
it is stuck on version `0.5.4` and hasn't been updated in
10 months. Instead, there is `@vitest/eslint-plugin` which has
the same author and same readme, but is up-to-date, being last
updated 5 days ago (as of writing).

There is only 2 warnings left on:
- `@typescript-eslint/utils`, but that just seems to be an
issue on yarn's side as it does exist in the yarn.lock and is not
something we need to add to the package.json.
- `eslint-plugin-storybook@0.11.3`: this has not been updated to
allow for `typescript>=5.8.0`, but there is little we can do about
that.

I also moved `react-icons` from a `peerDependency` to a normal
`dependency`, as per the comment in the other PR.