feat(project): Expand README and add settings.json
I have updated the README to include more information on project
guidance, like styling of storybook files, references to use, etc.
This is helpful for anyone working on the project.

I have also added the settings.json file from the DRL, which will
help configure the project for each person.