fix(project): Add necessary dependencies
I've added dependencies used by this library so that we can have
next-based components in the library e.g. custom `next/link` component.

I also added some vite plugins to manage static image we may want
to provide as well as SVG management. It's not as simple as adding
images to the repo, unless they are only used internally, if we want
to be able to import them into app code, then they need to these
plugins.