fix(project): Ensure `use client` and `use server` are honoured
Once of the big changes I've made is this, before vite/rollup was
stripping out this code, so that when I used the component in an
app, it complained that it needed `'use client';`, for example. I
couldn't understand at first and only with the help of ChaGPT did
I figure out it was vite/rollup. So now these lines should be preserved
in any module that uses them.