feat(project): Run script after dependency installation
There are some manual commands needed for the repo, to manage a local
GIT message on commits and setting up husky hooks etc. Instead of
relying on devs remembering, I've wrapped these in a shell script and
call it from the `postinstall` life-cycle script in package.json.
This ensures the commands are run after dependencies installed, which
is generally one of the first things a dev does.